Refactoring ProductsController - método crud e específicos

+++ ACMEStoreWebAPI/Models/Product.cs
@@ -7,12 +7,12 @@ namespace ACMEStoreWebAPI.Models
 {
     public class Product
     {
-        public long id { get; set; }
+        public Int32 id { get; set; }
         public String name { get; set; }
         public String description { get; set; }
         public String pictureUrl { get; set; }
         public Double unitPrice { get; set; }
         public String status { get; set; }
-        public long ownerId { get; set; }
+        public Int32 ownerId { get; set; }
     }
 }
